Add files via upload
This document is about Mining Association Rule. Mining association rules is a technique used in data mining to discover interesting relationships, associations, or patterns among a set of items in transactional or relational databases. This approach is particularly well-known for its use in market basket analysis, where it helps to find regularities in the shopping behavior of customers by analyzing large sets of transaction data.